Allow multiple column selections
Before this change the user could only select a single column in
the translation editor.

Start keeping track of the column selection in FrozenColumnTable
since the selection can span over 2 tables.

Override keyboard and mouse navigation to use gotoColumn() which
should:
- update the column selection in both tables
- make sure to `fireSelectedCellChanged` to update the panel
- reset the row selection (to single row) if not extending

Override keyboard navigation to use geteRow() which should:
- reset the column selection (to single column) if not extending

The StringCellRenderer is modified to paint a cell using focused
colors if one of the tables have focus.

Fixes: 117485022
Test: FrozenColumnTableTest
Change-Id: I011f40ccd8fe9e5bb2554a852c67a0cb783ad7af

GitOrigin-RevId: 2234907c80324d28195de8f23d253dd772177a8b