Replace AndroidDispatcher.workerThread with
... Dispatcher.Default and inject dispatchers.

This CL also makes the coroutine dispatchers class fields to make it
easier to inject the scope from the test in the future.

This CL also moves the getDeviceID into the WearPairingManager class in
order to access the backgroundDispatcher field.

Finally, this CL replaces the use of the IO dispatcher with the default
dispatcher for the loop that periodically updates the device pairing
states.

Bug: N/A
Test: existing
Change-Id: Ida5c54a1eca87cdff9901272bb9b8707c792b491

GitOrigin-RevId: 079828fbd610b0d34e5e4b8ca301bb21577e8f65