Fetch docs from contentserving API
This change lets us fetch documentation directly from the
contentserving API instead of from the DAC webserver. This approach
substantially reduces payload size and in turn, download time.
The feature is behind a flag and tests are added for both possible
values of the flag. Note that because the contentserving API
returns "SafeHtml", it means we need different content to compare
against, as it insists tags are closed, etc, in the HTML.

Bug: 391444348
Test: AndroidSdkDocumentationTargetProviderTest.kt
Change-Id: I758da16e7ca9660f0fa9654c335e9d7ed5bb7f2d

GitOrigin-RevId: feb14a18013b0c8355f647a9bd2e87e387f5498c