Move AndroidLintExifInterfaceInspection to new interface
Also simplify (and arguably make more correct) the logic around which
version of which artifact to register if necessary.  Determine whether
to use the androidx artifact not by whether we can find it on maven,
but on whether the module is configured to be part of an androidx
project or not.  This changes the behaviour in the
DefaultProjectSystem, unused in production but used for tests.

Also remove the complexity around assuring that the version of the
support library exifinterface artifact, if that's the one we're using,
is at least 25.1.0: no-one has been publishing let alone suggesting
the use of com.android.support:exifinterface versions less than 25.1.0
for quite some time, and it would be pretty weird to be in a situation
where we trigger this inspection and no exifinterface versions
satisfying this constraint are available.

Bug: 279886738
Test: manual, existing test adjusted
Change-Id: I7eacf94c950863faa8033098df972a1ab597162c

GitOrigin-RevId: d7ca8abf0c21783b28fd8150853c686f677a94fd