Update AehdWizardTest.kt to work with headless dialog change
When constructing the FakeUI for the old wizard tests, it will fail if
the root element is a Window element. This is currently not the case,
but there are changes to HeadlessDialogTester that break this,
ag/32231949.

Updated the old wizard test to match the pattern used by the new wizard
test when constructing the FakeUI. The dialog still doesn't have a
Window as the root component once it has been shown with `show()` so
the assumption holds again.

Bug: 402452919
Test: AehdWizardTest.kt
Change-Id: Ifaf725ccb8ce270a885b010328cfd19dfeafa4ce

GitOrigin-RevId: e69a47e5438b97a4ab9328c7a5521aa170db9cd4