Wait for indexes in one test
My working hypothesis for this is that the other tests in this file,
operating as they do through tempDirFixture, do whatever is necessary
to have the Vfs and indexes in a consistent state.  `writeToBuildFile`
bypasses some of that: in fact it looks quite different from
CodeInsightTestFixtureImpl#copyFileToProject which has a similar end
state, missing out document reload and index waiting.

Adding those missing parts to GradleFileModelTestCase#writeToBuildFile
and friends might be a more principled way to address this, but causes
a ~30% slowdown in the execution of this target (and presumably
others), probably because it undoes the optimization of skipping the
initial indexing pass in the first place.  Inserting the wait for
indices in this single place de-flakes the test with minimal effect on
test execution time.

Fixes: 400927524
Test: re-enabled
Change-Id: I620904ab32fc2181d3916b954fc19ac8a0e497ac

GitOrigin-RevId: 7867e2964a56dd3c768aab0570ddaf7910e526e7