Fix SceneViewPeerPanel memory leak
The LabelPanel objects use their coroutine scope as disposable parent
for a message bus connection. However, the coroutine scope comes from
the creation of SceneViewPeerPanel and is derived from the one from
SceneViewPanel. Since SceneViewPanel is a much longer lived object than
SceneViewPeerPanel, the derived coroutine should use SceneView as a
parent disposable to ensure that it gets cancelled when the
SceneViewPeerPanel gets destroyed.

Bug: 401823931
Test: N/A
Change-Id: Ibacf89f76c556f6c4b35c2a3100605743f9937e5

GitOrigin-RevId: d026815088b1253566e03f3af162937926987d77