Make image a property of VirtualDevice
The original reason for image being separate was that we didn't have
nullable properties on VirtualDevice (VirtualDevice always had to be
structurally valid), and there was no suitable default. Now that
VirtualDevice can represent incompletely-defined AVDs, we can simplify
the code by putting image with all of the other AVD properties.

This will make it simpler to validate device properties like RAM, for
which the minimum depends on the chosen system image.

Bug: 382095105
Test: updated
Change-Id: If28f06ecb2ff928a347069b867a5b2f57fd2e032

GitOrigin-RevId: d7fb599ff9c6c79b482bf028c3bde309bd2331e5