Introduce a new test rule
cl/735880054 introduces _wrapped_android_local_test. It should be
treated as android_local_test and should be debuggable. At the same
time, android_local_test rule is still used, so we cannot replace it.

This is google3 only but it should affect bazel users who will not get
this rule.

Bug: b/404454900
Test: manual verify new test rule debuggable, we may not be able to add
 test for blaze only feature as all repo test are run with bazel

Change-Id: Ib91fff0c32f24e534f3173906236148fb039d85f

GitOrigin-RevId: f9cd2214a2f9a59af1ead8897432450465a28fb8