Remove deprecated usages of Query.asIterable()
Query.asIterable() is deprecated in favor of APIs that make clear that
the entire collection is generated eagerly, rather than lazily on
demand.

https://github.com/JetBrains/android/commit/ed0bc6b860a78c1c0ce1d4fa80d5286839f785fd
made these calls explicit (rather than implicitly using `Query` as an
`Iterable`) so they could be more easily identified and removed. This
change replaces the calls with a more appropriate alternative.

Bug: n/a
Test: Covered by existing
Change-Id: I16e7b6f8fa7d182525bf937a4084602b8619a96f

GitOrigin-RevId: a11ed9a75135636f7d24af9f57b673259e10b2a5