Update OutputArtifactInfo's way to get relative paths
Currently the method returns the path that is relative only to bazel-out
directory. This change makes it more generic since that is needed for
deploy jar paths for debugging local tests.

This change should be safe since it doesn't change the behavior of the
getBazelOutRelativePath() method, which is used widely in the code. It
only changes the behavior of artifactPathToBazelOutRelativePath().

Bug:399949720

Test: n/a
Change-Id: Ic342b57ca099c37542eca49f83c1dca7d325da55

GitOrigin-RevId: 9aea53d746a7d36b2cc6d5d46869e407a857bee4