Specialized support for LintProject#dependsOn(String)
There are two cases that we actually support in AndroidLintIdeProject.
The first is a dependency check for a google maven artifact, which
stands a chance of working in all deployment scenarios; the
second is a check for an arbitrary maven group:artifact identifier,
which only really stands a chance of working when the underlying build
system is Gradle.

In AndroidLintIdeProject, support these two scenarios explicitly,
rather than implicitly through a supposedly-general but in fact
specific AndroidModuleSystem interface: if we can find a gmaven
artifact identifier, use that; otherwise if we have a
GradleModuleSystem, use the more specific facilities offered there;
otherwise, delegate to the superclass method.

Bug: none filed
Test: existing
Change-Id: I30f0b6d2059f4f7b75d73374719ec51619606885

GitOrigin-RevId: c12134ca1a388729c431920d61fc859e782b3070