Make tooltip follow cursor
Labels in VitalsIssueDetailsPanel can get long due to the device name
and date format. The default behavior on the label is to show a tooltip
where the cursor first entered the label. Given the length of the label,
the entry point and the location of the cursor might be far enough to
make it look like it is a misplaced tooltip.
This change hides the tooltip whenever the cursor is moved forcing a
redraw of the tooltip closer to the cursor.

Bug: 382414220
Test: Manual
Change-Id: I972f3834d373ee2036eb9213f0373336867dac84

GitOrigin-RevId: dd0252c3d04369353ca0924acad05da75d2fb9c9