Fix flakiness of testAppearanceAndToolbarActions on Windows
There were two causes of flakiness. One was an incomplete update of ActionButtons in FakeUi.doUpdateToolbars. Another was an asyncronous high quality repaining of the display image using a more acccurate scaling algorithm. On Windows, due to slowness of the the build machines, we sometimes got high quality rendering that didn't match the goldens.

Fixes: 403316901
Test: DeviceToolWindowPanelTest, 1000 runs on Windows without flakes
Change-Id: Ic11c0b401959cff2c0be1897c4286271e8e163f0

GitOrigin-RevId: 27a5632dbf21577e24f98dd8f00bbae5477267a9