Fix AndroidCoroutineScope test
The test framework for coroutines is set up to fail if any Jobs don't
complete successfully. Given that this test requires cancelling a job,
it was running into problems.

By using EmptyCoroutineContext instead of inheriting the test context,
we can still test that the Job is cancelled on Dispose of its parent
without failing the whole test.

Also adding some other minor cleanup of deprecated items.

Fixes: 303086924
Test: androidCoroutineScopeIsCancelledOnDisposeInRunBlockingTest
Change-Id: I0004da9372f4a2a8d9dd56f0e574e3f88504e730

GitOrigin-RevId: ac8bad9dd1b74d7338b119b26852386b3ca9a311