Use newly created GoogleMavenRepositoryV2 in RepositoryUrlManager.
This change introduces a studio flag which can enable GoogleMavenRepositoryV2. Currently, the main logic is not introduced but this is part of a wider effort to refactor GoogleMavenRepository to hide the implementation details from the callers. Next, we will introduce the implementations for the methods used in "RepositoryUrlManager" and enable it for Canary for a phased rollout.

More details in go/studio-template-dependency-impl

Bug: b/407727269
Test: Unit tests and WAI screenshots in comments
Change-Id: I79b71345b41dcb66808fcc20cbb59e5fac1314cc

GitOrigin-RevId: cafc23743014b6679b3f97114903e087b12064d8