Fix StudioDownloader progress reporting
There are two interfaces for progress reporting: the SDK's
ProgressIndicator and IntelliJ's ProgressIndicator.

StudioDownloader accepts an SDK ProgressIndicator on its various
download methods, with which it instantiates a
DownloadProgressIndicator. DownloadProgressIndicator also grabs the
current IJ ProgressIndicator, and sends its updates there as well.

However, the passed-in SDK ProgressIndicator already delegates to the IJ
ProgressIndicator. It also has a hierarchy of indicators created by
createSubProgress, so that individual tasks report their own progress
independently as part of the overall progress.

This means that IJ receives updates twice: once directly from
DownloadProgressIndicator, and again, but with an adjusted scale, from
SdkManagerProgressIndicatorIntegration. This causes the progress bar to
jump back and forth constantly.

Instead, we should just let updates propagate linearly:
StudioProgressIndicatorAdapter is now just an adapter from IJ to SDK
ProgressIndicator. The other clients of StudioDownloader
(HttpArtifactResolver, CpuTraceInterceptCommandHandler) are not
expecting updates to go to the UI.

Also, createSubProgress needs to read the current fraction in order to
set up its scale; thus, we need to make sure that setFraction and
getFraction are implemented properly in the various progress plumbing
classes.

Also, use sub-progess within StudioDownloader to avoid jumping
backwards, and use try-with-resources on our InputStreams.

Fixes: 64794823
Test: Updated
Change-Id: I74e19a7326bdd2e5b1e091281cd2a578326bd63f

GitOrigin-RevId: 879011e3408cc31f5e67e66c646436df7cd90699