Fix color picker failing to update
The call to `myElement.getProject` is sometimes failing with a coroutine
cancellation exception. This appears to be a relatively recent change,
as the platform infrastructure for showing the color picker was
rewritten with coroutines.

I'm not entirely sure how this can happen, since the call is coming in
on the EDT and I don't expect it to have been cancelled. Regardless,
it's easy to fix by storing the reference to the project earlier and
using it now to schedule a later action. This also shouldn't have any
dispoable reference issues, since the class is already storing a
reference to the PsiElement (which itself contains a reference to the
project).

Fixes: 406196949
Test: Manually verified
Change-Id: Iea391dfbbf758bd9dc10703fb9db129adb83206b

GitOrigin-RevId: 07cbf896e63d1bc8144c902d892252d944b1b675