Fix race condition in GradleFiles update
Since we now check for presence in the hash table to know whether we
care about Psi events coming into the GradleFiles listener, we must
only schedule recalculation of the hashes at the end of sync, rather
than at the start, because the structure of the project might change
as a result of sync (e.g. bringing in new Modules, and hence new build
files).

However, doing this reveals another issue, which was already present
in the previous code, but latent: in finding build files for
modules (and the Gradle wrapper file), we implicitly request VFS
refresh, which we must not do under a read lock.  Move code around so
that we are not in any danger of doing so, and modify the way we wait
for completion in unit test mode so that we don't actually cause a
deadlock.

Bug: 406805973
Bug: 406806595
Test: caught by existing UI tests
Change-Id: Iaf96cbe54827723e9ef2555c30a4314e0c582fe7

GitOrigin-RevId: 223580c0d72ef7b65bbfaa0df436556af3976f29