Allow hiding validation message panel when empty
The validation message panel, used by the ValidationPanel component to
display error and warning messages, is currently always visible. This
can lead to unnecessary visual clutter when no messages are present.

This commit introduces an option to configure the ValidationPanel to
automatically hide the message panel when it has no messages to display.
The visibility of the panel will be toggled dynamically based on the
presence of messages.

The option has been enabled and tested in the following wizards:
* 'New Android Fragment' wizard
* 'New Android Activity' wizard
* 'Configure Image Asset - Asset Studio' wizard

Fixes: 286593515
Test: ValidatorPanelTest + manual checks
Change-Id: I629ed093ee06b7ccf9a452ba28a11665e051f115

GitOrigin-RevId: f9788b436318fec06357f0fe2235e7fed6f3f683