Track all running emulators including not launched by Studio
LaunchedAvdTracker has been extended to all running emulators and renamed to RunningAvdTracker. This class is intended to be used by the Device Provisioner and as a source of information about running emulators that are not yet or no longer connected to ADB.

It keeps track of all running emulators including the ones started by other processes. It now knows when an emulator is being shut down by the EmulatorController.shutdown method. The isShuttingDown property in not yet updated when an emulator is being shut down by the Device Provisioner. In order for it to happen Device Provisioner should call the RunningAvdTracker.shuttingDown method too.

Bug: 339688293
Test: RunningAvdTrackerTest
Change-Id: Ice2d6c62ec6ac79d636204e7c4e71d6c31f8f2ea

GitOrigin-RevId: 66b73de9911a2fc0b8ea80a0b65d103c8b4da8c1