Stop tracking build directory for snapshots tests
The ProjectDumper tracks the existence of the different paths by
adding '[-]' as a suffix if file doesn't exist. Despite this being
useful to understand if certain files exists creates unnecessary
noise with build directory, since the presence of it depends on the
implementations of different KGP, AGP and other project/dependencies
versions, which might end up generating files or not during sync.

For this reason the build directory will be ignored, however, expected
generated files will still be tracked. This last point is something we
would potentially revisit in the future considering to exclude those
that doesn't add any value given that expected AGP generated file are
already validated separately.

Bug: 405366819
Fixes: 403189695
Test: Existing
Change-Id: I595f23e7d4efc699e7b65fcbb131e883eda1acdf

GitOrigin-RevId: ec6e62f0229b86351531961d20ef81ba8de77c36