[SDK Index] Show Deprecation issues first on PSD
... when different errors are shown for the same library, this is needed
since the only way to remove the issues is by removing the dependency.
Other issues may contain quickfixes that only change the version and
thus it is important to say the dependency is deprecated first.

This change adds a priority to PsIssues so we can have high priority
issues within the same severity. The comparator now uses these values:

- Severity
- Path
- Priority
- Text

Bug: 383771175
Test: IssuesBySeverityPathTextComparatorTest
Change-Id: Idf0fae525d76ede01332b2b7f2cf034a36aa55d8

GitOrigin-RevId: 535023bafd9a27e2639045ccf0e08eabae576543