JPS: exclude resource files under java/kt roots
Now we collect resources from explicit resource roots only.
See b/405988245 for details.

Surprisingly, but conveniently, the 'wildcardResourcePatterns' setting
applies to source roots only, not to resource roots. This fact is not
documented in https://www.jetbrains.com/help/idea/compiler.html
but it is the observed behavior, and it is implemented this way in
org.jetbrains.jps.incremental.ResourcesTarget#computeRootDescriptors.
So, we can use pattern "!*" to exclude all resources in java/kt roots.

Bug: 405988245
Test: JPS build + launch Studio from the IDE
Change-Id: I338495b1a5288dd599f63338ba72c6c495db72ea

GitOrigin-RevId: 83ce23188f444d558dfba5fd72dc485c32eb91c2