Change mocking style in AndroidGradleProjectStartupActivityTest
The error message in the flaky test refers to this being the preferred
style of mocking for spies.  We don't obviously have a spy, but switch
to this style of mocking anyway.  (I guess it's more likely that this
flakes from the interaction between mocks and multithreading, but it's
worth doing the simple thing to see if it makes the problem go away.)

Bug: 380326797
Test: existing
Change-Id: Ieaa1fcf6e7e77d7542467a914ddb3e083a9de444

GitOrigin-RevId: a4dd241b62c14ee100fe8ab8388695a0349529ae