Fix NavDesignSurfaceZoomControlsTest
Test cases for NavDesignSurface has been affected by flakiness for the
following reason:
1. NavSceneManager was accessing a psiTree on EDT
and it needed to be wrapped in a runReadAction first.
3. Not searching the diff image in the right time of the rendering.

Fixes applied in this CL:
1. adding a `runReadAction` in NavSceneManager.
2. re-writing NavDesignSurfaceZoomControlTest.zooOut and
NavDesignSurface.zoomToFit test cases in a way that the diff images
are checked until the rendered image is equal to the golden image. And
also removing the deprecated Dispatchers.

Fixes: 408152629
Fixes: 402505835
Test: NavDesignSurfaceZoomControlsTest

Change-Id: I4986dfbb440a56e3c274a11365c293fecc9b6b08

GitOrigin-RevId: f8c253bc5a5ff69c0daa0472ddb7b22861dc8b8e