Postpone reformatting slightly less
When applying changes to the project build model as a whole, don't
delay formatting around the entire call, because that might (depending
on the details of the postprocessing) not leave the individual files,
documents and Psi in their expected state, either at the point of
file.saveAllChanges() or at the end.

I couldn't find a way of triggering any failure case, but it looked
suspicious while debugging, and probably doesn't give noticeable UI
freeze wins (compared with the inner postponement around a single
file, which almost certainly does.

Bug: none filed
Test: existing
Change-Id: I1d2fe43728aea6f5abf666831b7bb34a8f6185fb

GitOrigin-RevId: 5315e77df0348537a0e419b617dc7b6ce9eab36f