Order operations in tests more carefully
We must not create a NewBaselineProfilesModuleModel for a project
before having loaded the project, because some of the construction of
the model involves attempting to find out properties of the project
itself (for example, whether view binding is enabled).  This in turn
potentially constructs and caches an instance of the
DefaultProjectSystem, which will later give wrong answers to other
questions.

Bug: none filed
Test: test-only change
Change-Id: I8b305be3831673608babf10cc9b39df452471521

GitOrigin-RevId: 2526733b21f961b4b856f110373a1cde8e59f087