Snapshot changes for phased sync (composite builds)
This is mostly renaming the java project modules, and making it more
consistent with the Android naming. Formerly, for some reason, the names
were derived from the project group/version (from the build script), and
now they are not, due to sync not having that information yet when the
modules are being created. Technically we could rename those to keep the
old behaviour, but I deem this change safe, so I kept it.

There are also changes in some paths in the compiler module extension,
but those changes are only in the module corresponding to the Gradle
project, not the source set. And the output should be dictated by Gradle
anyway so those changes should be safe.

Bug: 384022658
Test: snapshot change
Change-Id: I7b20d46ea58e6499b6e946ab113f2ac7c7ba7d14

GitOrigin-RevId: 4a3cbc9d59f262d1f06647b103300aed14358ad8