Ensure a refresh request succeeds
The CommonPreviewRepresentationTest was flaky due to a condition
that sometimes didn't fulfill properly. This is caused by a race
condition that sometimes leads to a refresh being "skipped" due to the
rendering build status not being ready.

This CL ensures we have a successful refresh request by manually waiting
for the expected rendering build status.

It's a temporary measure to fix the flakiness, however the race
condition itself is not fixed.

Fixes: 409371897
Fixes: 390629503
Bug: 383310211
Test: bazel test --config=remote //tools/adt/idea/preview-designer:\
  intellij.android.preview-designer.tests_tests__other \
  --test_filter=CommonPreviewRepresentationTest --runs_per_test=1000
Change-Id: If37da3feab9b435aaf5af626b33ef69ed270593e

GitOrigin-RevId: 08a12b4173016e2a7580abede539258d5e9f4ce2