Fix waiting for extract function
I previously introduced a semaphore to wait for the extract function
action to complete, but that sometimes doesn't get triggered correctly
and causes a hang when there's an error.

Waiting for a non-blocking async read action appears to be the "correct"
way to handle this, as it's what the Kotlin plugin is doing in its own
tests.

These tests are still flaky, but now the failure actually reflects the
real exception, not a timeout on waiting for completion.

Bug: 408008767
Test: ComposableFunctionExtractableAnalyserTest
Change-Id: I19a5412f00fc713093ad872e3f9b614c7c22714d

GitOrigin-RevId: 33663cb692f4c9f8ddf4e1051ef9399c6dab64ee