Fix test flakiness
This test just needs to wait for indexing to complete. I had tried
something similar before, but the previous mechanism in place for
waiting (with a semaphore) put a gap between the EDT action and the
background action during which indexing could start in some cases. That
no longer exists, and this test passes in a 700x run:

https://fusion2.corp.google.com/invocations/349e9d8e-44ae-4e82-840a-1d78c9058094

Fixes: 408008767
Test: ComposableFunctionExtractableAnalyserTest
Change-Id: I271a72f125d2a5c859bd861ec2b7e6d5145865a7

GitOrigin-RevId: 9ba7b3d9ee3eba609a235861132012c01f1afe46