Ask Gemini for context file names
This changes how AQI gathers context. It now prompts Gemini for a comma
separated list of files that it needs, resolves the content of those
files from the project system, and then prompts Gemini again, this time
for the insight.

Introduced a new flag to gate this behavior. Moved the call to
CodeContextResolver to GeminiAiInsightClient. And a bunch of other minor
refactoring changes.

Test: new and updated tests.
Bug: 408451958
Change-Id: I2e3e7e30737a1cfda0812f9b843317ebc01db0c4

GitOrigin-RevId: 9b7eb21be5bed7a5fc16e55ebb5ab3f8e3c8f522