Override and use different getQuickFixes
Fixes are created whenever the previous getQuickFixes (the override
with the FixData parameter) method returns an empty list and when the
FixData is not null.

Because we were returning a QuickFix from that method, the other fixes
that were defined in the Detector were no longer being created. This
was fine when we were only reporting the missing property issue.

However, we now want to declare other issues with fixes defined in the
detector. We can do that by overriding a getQuickFixes with a
different signature.

Bug: 406968509
Test: manually verified
Change-Id: I433cf83429cf10f18c4f609172a56745bb077b51

GitOrigin-RevId: c3135eea1dabc03c43d2bdb3f2b3eff5ff8a8d6d