Fix flaky test
The flakiness was caused by kotlin classes not being present in the
PSI class index. Because of this, we were not creating an expected
number of run configurations. The issue only happens on Windows.

I was not able to reproduce this issue when creating
an isolated test and when running the flaky test case in an isolated
environment. This leads me to believe that the flakiness is not caused
by a platform issue but rather by the test infrastructure.

The CL fixes the flakiness by converting the kotlin classes into java
to ensure they are present in the index when creating the run
configurations.

Fixes: 410256906
Test: c:\src\studio-main\tools\base\bazel\bazel.cmd test  \
  --runs_per_test=1000 --config=remote \
  //tools/adt/idea/android/project:intellij.android.core.project_tests \
  --test_filter=AndroidRunConfigurationsTest
Change-Id: I8e0f4ac25a29361428cd55184f463227b61c7488

GitOrigin-RevId: b3a1493889c285f6ebdb912099a990a6320d3f0a