Add OWNERS to android/gradle
This module is now test-only.  There is some conflation in the tests
between those that actually test parts of what is now the Gradle
project system, and those that simply require a full Gradle sync to
function, but it's probably not wrong for tests that fail in this
target to be routed to the gradle project systems team in the first
instance, and also that that team should have oversight on changes to
the tests.

Bug: none filed
Test: n/a
Change-Id: I445360d83a6e51876dfcefbd12327d8afded2725

GitOrigin-RevId: 933869e7a3246769514453124527c371791e5830