Improve JDK assertion for the Gradle Daemon execution to sync project
Before Daemon JVM criteria, the IDE was responsible to indicate to
TAPI via LongRunningOperation the java home to run the build/sync,
for this reason resolving the project Gradle JDK configuration was
sufficient.

However, that will not work for projects using Daemon toolchain since
Gradle is responsible to auto-detect or auto-provisioning a matching
toolchain for the specified JVM criteria.

To address this the existing JDK assertion for JdkIntegrationTest
now will use IdeaProject Gradle model instead, containing resolved
JDK used to set up Daemon JVM to run the sync.

Bug: 395610076
Test: MultipleGradleRootSyncUseDaemonJvmCriteriaIntegrationTest
Change-Id: I7b946c7c7e181cf60760168f70fca9cba341c705

GitOrigin-RevId: c8be2795f0c3215fcef78110fbff02ee39b37592