[Android View] Add settings to make Project View default
This change adds a new settings tab to customize the Project tool window
in Android Studio. The only setting added at the moment is to allow
users to set the project view as default. On previous versions of Studio,
users can set the property "studio.projectview=true" to change the
default. After this change, this setting will be stored as a
persistent setting but will continue to take the property value.

The UI can have 3 states now:

- SHOW_DEFAULT_PROJECT_VIEW_SETTINGS not set
  - Setting is not shown
- SHOW_DEFAULT_PROJECT_VIEW_SETTINGS set and studio.projectview not set
  - Setting is shown and enabled, user can make changes
- SHOW_DEFAULT_PROJECT_VIEW_SETTINGS set and studio.projectview set
  - Setting is shown but not enabled, user cannot make changes and the
    message reflects that the property is set.

Bug: 399814788
Test: AndroidProjectViewTest
Change-Id: If7b91d23d642c4aa07cbba4d6ba12cb104ed31d7

GitOrigin-RevId: 0eba06f5478dfa598b2d5285a7d72dcb8f4e8fe6