[gradle][android] fix test AndroidBuildScriptsGroupNodeTest.appProject
This test was failing on line 92 `assertFalse(node.contains(app_gradle_libs_versions_toml.virtualFile))`. It checks that the app/gradle/libs.versions.toml file is not considered as a version catalog.

I assume that in this test case `app` is a subproject, so I included it in the root settings file. Without this change, the root build's settings file and version catalogs were not used for the `app`. So it was considered like a separate build that could have its own settings and version catalogs. See `BuildModelContext#canBuildUseSettingsModel`

GitOrigin-RevId: c484c6350c5c93241a14f1f4d052fbac70e39218