Prevent layout inspector from crashing when closing tabs
This issue happens because the emulator takes an extended period to
disconnect.

Consequently, when a tab is closed and the layout
inspector is restarted on the previous tab,
the soon-to-be-disconnected emulator is still considered a
connected stream, leading to a crash.

To address this, a condition has been added to ignore events
not originating from the currently selected device
(if one is selected).

It still throws some warning due to streams got garbage collected
without being shutdown first. I will issue a separate bug for it.

Bug: 406777830
Test: ForegroundProcessDetectionTest
Change-Id: Id5d3aa5b38294df7ea603546c85ba092399e2bd8

GitOrigin-RevId: 5460f9b8b46c144ce421d120aef606be110bbc73