Make Room @Query Annotations Work With KMP
The existing code checks if the project includes Room by searching
the for Room classes in a  `GlobalSearchScope`. The scope used is
the `module-with-deps-and-libs` scope but it seems that in a KMP
project, the libraries are not properly included in the scope and
we fail to detect that Room is included.

Looking at the scope returned from
`module.getModuleWithDependenciesAndLibrariesScope()`, `getRoots()`
returns `klib` files instead of `jar` files:

```
  {FsRoot@72080} "jar:///home/aalbert/AndroidStudioProjects/Fruitties/.kotlin/metadata/kotlinTransformedMetadataLibraries/androidx.room-room-common-2.7.0-alpha07-commonMain-gMSPmw.klib!/" -> {Integer@72081} 29
```

Bug: 379001138
Test: Manually
Change-Id: Icf2bb4c64be5e4355eae212dcb793df0a8fd691a

GitOrigin-RevId: a168c3c1c3d1caa2bbba0c9e11648dcbfd22d1c8