Clean layoutlib native memory when exceeding threshold
This adds an additional step in the RenderTask dispose that attempts to
clean up native memory used by layoutlib when it is above a certain
threshold. This uses one half of the JVM heap size as a threshold to
ensure a high-enough threshold.
This tracks when the threshold gets breached using Studio metrics.

Bug: 401823931
Test: N/A
Change-Id: I95f280b5351a0ff6c11db3914f013ba70d3dcb64

GitOrigin-RevId: a6e62e7c7d516fcee11ca36d484b3fd2246075ed