Use isBazelProject in BlazeRenderErrorContributor
and a delete negative test that only tests applicability to an invalid
but still bazel project. `isBazelProject()` is still an incorrect and
deprecated way of checking whether a project is a Bazel project.

A better way would be to check its project system, but this is not
yet possible because of the current project structure (base vs. aswb).

Bug: n/a
Test: n/a
Change-Id: I25606907506b829d6b9b98b7a65f67c3566543a3

GitOrigin-RevId: fcb0191f1596be0e68a1c569339809f33079c42a