[Android View] Show build scripts in module
... when ProjectToolWindowSettings#showBuildFilesInModule is true.

We currently show all the gradle build scripts in a single group node,
but for large projects this can be cumbersome since it is harder to look
for a particular file this way. This change displays build files on its
module as opposed of a single group node when showBuildFilesInModule is
set.

Bug: 298438420
Test: AndroidBuildScriptsGroupNodeTest
Change-Id: Icb83cc2c7c41575d8fbae82dc0d683399f745a10

GitOrigin-RevId: a2af6c8b5426a1bb258a71d63ef670dade3a68f2