Implement a Component-based interface to the dependency analyzer
This allows answering the question about whether a set of Gradle
Components is compatible with the existing dependency
declarations (and each other).  Previously this was done by the
GradleCoordinate method, but we are choosing to make the distinction
between Components and Dependency declarations clearer, and the
GradleCoordinate entry point is going away.

Convert most tests to the Dependency entry point, which does no
stability analysis on its arguments, taking the view that a
Dependency's rich version already expresses any stability that the
user wishes to impute; implement an additional test to verify the
Component-based behaviour.

Bug: 279149611
Test: yes
Change-Id: Id1983930e88288bff730276fe9f55ecd72cb1110

GitOrigin-RevId: 440b22a6d5df458c82ece5720511e65bde083e0f