Adjust tests for fixed RichVersion printing
The Gradle Dependency Compatibility Analyzer converts requests
involving GradleCoordinates to Dependency specifications with strict
version requirements, in order to be able to model known
incompatibilities, for historical reasons.  Until we corrected the
printing of RichVersion with a strictly declaration but no preferred,
this was mostly invisible; the conversion from a specified version to
a range was perhaps surprising, but not so surprising as additions of
two exclamation marks.

Bug: none filed
Test: adjusted
Change-Id: I08f0d9cb78e32bcc466293c6751ecdb8ff370c7f

GitOrigin-RevId: ef136226adf25ca9a12eec3c7f7359323163a527