Postpone selection change listeners to after update is completed
When detecting a change, there is a chance that the
TreeTableSelectionModelImpl is in the middle of an update. At the moment
the code will ignore that selection change. This will cause the nav
editor to miss some selection listener changes.

This ensures that, if the model is in the middle of an update, we
schedule the notification listener to immediately after.

Bug: N/A
Test: NavComponentTreeTest
Change-Id: I1c705dba24b0e13c11b627942b66bdd3f69edaf3

GitOrigin-RevId: 2ea79d1407878a6187c2d8d2242f96c6df4ae1b0