Update AndroidVersionsInfo for minor versions
Stop using hash strings in buildApiLevelStr for platform targets (i.e.
"android-35"); these are not actually accepted. (Retain it, for now, for
add-ons; these also don't work, but add-ons are filtered out later.)

Use apiStringWithExtension for buildApiLevelStr; this means that minor
versions are included. (They don't yet make it all the way to the
rendered templates, but the data is at least plumbed to
ApiTemplateData.)

Use featureLevel consistently for buildApiLevel and minApiLevel.

Replace various deprecated methods with their equivalents.

Update test to expect minor versions and not expect "android-" prefixes.

Bug: 398938512
Test: AndroidVersionsInfoTest
Change-Id: Ia60421d804f796996f48db81af628660c8903c89

GitOrigin-RevId: 02a44e7a61348b42f604a38104178552cfec00bb