Provide more models in PhasedSyncProjectModelProvider
There seems to be an issue in some older AGP versions where fetching the
AndroidProject causes the subsequent fetch of BasicAndroidProject in
full sync to behave differently for aidl/renderscript directories. So
fetching of any model except Versions is disabled until 8.0. We might
reconsider this if we ever implement a cache across model providers
to avoid fetching the same model twice.

Bug: 384022658
Test: Existing snapshot tests
Change-Id: Ib70961068e713cd3737f2a58688288b07d73f91c

GitOrigin-RevId: 2b9da707f4cb8ae31b667fa1ba1d6e5bbc4053e2