Populate variant specific source sets in phased sync
Normally to correctly pick the variant for each module, we rely on
dependency resolution but with phased sync, we can't do that. So instead
we're picking the variant using the following priority:
- User-selected variant for a specific module (through IDE UI)
- Previously selected variant (stored in facet)
- Default variant, as denoted by the build script

This also implicitly drops support for KMP projects (by filtering
modules that end up with duplicate content roots) as the initial
support was incidental and very basic. For KMP projects the way to go
should probably be to have their own sync contributors anyway.

Also accordingly updated the list of known issues in tests and had to
add some more, either due to surfaced new issues that were masked by the
parent "content entry" issue, or issues newly introduced due to the new
behaviour.

Bug: 384022658
Test: Covered by existing tests
Change-Id: I66392c03fc52d56d3eddd949c9f76eb7ebb53b5d

GitOrigin-RevId: 3393571b52e8c685dc8d7bdfbca2ee5aa5d299a2