Hide Query-Sync opt out settings in UI when auto-conversion experiment is enabled
When the Query-Sync auto-conversion experiment is enabled then the new
projects may only be created as Query-Sync (and not as Legacy sync). So
this fix disables the Query-Sync opt out settings in the UI when the
experiment is enabled. It also ignores the opt-out in the new project
wizard when the experiment is enabled.

Bug: 405438522
Test: Manual
Change-Id: Idd808f417918d6eff4e7c39e6240c31790dd4e06

GitOrigin-RevId: 8eeb761bf1e3860a89bbd4a0a4af5a1125f35c81