Make Hardware Input in XR a submode of Interact with Apps
This change replaces "Hardware Input" in the floating XR toolbar with a more generic "Interact with Apps". "Hardware Input" becomes a submode of "Interact with Apps" and treated similar to how it's treated for non-XR devices. This allows keyboard shortcuts to be used for switching between all XR input modes.

 It also opens a future opportunity to allow WASDQE keys to be used for navigation while holding some modifier key without explicitly swtching to a navigation mode.

Bug: 416893561
Test: existing
Change-Id: I926c7d6fd25f95e73e46d3ce0cc209bb214bb280

GitOrigin-RevId: 5e20a41eea23cb66c22ab3e2918560be4d3ded48