Minor modification to AGP/Studio compatibility computation
Treat the GRADLE_PLUGIN_MINIMUM_VERSION and
GRADLE_PLUGIN_NEXT_MINIMUM_VERSION as referring to release series,
rather than a specific version: so that the boundary for considering a
version supported is the first alpha, not the released version.

In practice this doesn't extend the supported window because a preview
version will end up with a DIFFERENT_PREVIEW state; it only makes
reasoning about which case will result somewhat simpler.

Bug: none filed
Test: added
Change-Id: Id424d3703a2bfbff2c8178b72a3deea49d315728

GitOrigin-RevId: 4b2d4493d4213165666b8acf099a1de3bda30f69