Update tests for new GRADLE_PLUGIN_NEXT_MINIMUM_VERSION
The changes to UpgradeAssistantContentManagerTest do not make the
tests completely well-described; there remains some conflation of "the
stable/deprecated version" and particular version numbers where we
test for the existence or display of of specific refactorings, but
it's better than before.

The changes in ComputeGradlePluginUpgradeStateTests should likewise be
read with caveats: as well as the obvious encoding of literal version
strings everywhere, the tests themselves hide an implicit assumption
that the support boundary for AGP versions is also a major version
boundary, which is true for now but might not be true for ever.

Bug: none filed
Test: updated
Change-Id: Iacf0a07fb6d1d656557fd7a4efada395fb24e261

GitOrigin-RevId: fcedf51a470a01cf9186fd4e6eb0b17d8b7324c9