Fix Round Wear devices are rendered square
When showing a round preview and switching from Ui Check and then back
again to Default and back to default, shape turns to not be round
anymore.
The bug is related with ui Check mode filter that shows a decoration
only if we are showing previews on Wear devices.

To fix this bug we allow UiCheckModeFilter to show decoration not
only when a wear device is defined but when overall the
PreviewDisplaySettings has showDecoration flag as true.

Bug: 409524854
Test: UiCheckModeFilterTest
Change-Id: I21eb5ca644a6df4033095ecf7d06bf2607600a33

GitOrigin-RevId: 31743327016ea51fa5389e0ddd6d928304312c92