Delete loadRemoteTargetVersions
It pretends to add versions to the list produced by loadLocalVersions,
but essentially does nothing.

addPackages tries to add API levels from available system images, but
only if they are 32-bit x86 (see getTag()). These are not produced
anymore (except possibly for TV).

addOfflineLevels adds KITKAT_WATCH back for phones (erroneously).
The rest of the code does nothing since we already have added the range
from minOfflineApiLevel to maxOfflineApiLevel in loadLocalVersions().
(Theoretically, we could skip calling loadLocalVersions() and only call
loadRemoteTargetVersions(); no callers do this.)

While we could fix addPackages, there is little reason for templates to
support setting minSdk to a version newer than what Studio was built
with. This would limit the app to running on the very newest
devices. It's more likely that the user would want to set compileSdk
to gain access to new features; however, we don't provide any
direct way to set compileSdk to a newer version than
NPW_COMPILE_SDK_VERSION. (Setting minSdk to a higher level is an
indirect way to force compileSdk higher.)

It would also mean that the user is targeting an SDK that is
unsupported by AGP, or running an AGP that is newer than Studio.
Upgrading Studio would be the recommended approach; experts can set
these values on their own.

Bug: 396502289
Test: AndroidVersionsInfoTest
Change-Id: I3b0af6f32647ada6db1c34ea3132400dc16ab426

GitOrigin-RevId: fb1d859f68156a760a7003dd8a09a3a47bc3888a