Do not rebuild dependency graph
unless the query result changes. Building dependencies do not change the
dependency graph, which is purely based on the query phase output. This
is not currently enforced by the shape of data structures but it should
become clear that no data flow from building dependencies to the build
graph when `QuerySyncProjectSnapshot` is restructured to reflect this.

It takes time to do rebuild it and with code analysis auto-enabled on
sync in might result in it being built twice.

Bug: 417935665
Test: n/a
Change-Id: Ia42ec10e106c3e8f28b06c12ae52b6967c242948

GitOrigin-RevId: 443ecf6247dda9f08287faa302932870bebbbdf2