Add fine-tuned const terms to performance test part2
In this cl we fine tune some of the [constTerm's](go/perfgate-wda#const-term) for the perfgate WindowDeviationAnalyzer that tracks regressions in studio performance tests. The WindowDeviationAnalyzer.MedianTolerance analyzer that we use uses the bellow formula to calculate the tolerance. From [here](go/perfgate-wda#tolerance)

  tolerance = const_term +
              median_coeff * abs(historic_median) +
              mad_coeff * historic_mad


In other words we are tuning the const_term part of the formula. The values there are around ~ 100ms-700ms for the metrics with values of 10sec-15sec (https://perfgate.googleplex.com/benchmark?benchmark_key=6097078101344256&~1e=median)

Bug: N/A
Test: StartupPerformanceTest
Change-Id: Id97f196e25f8f0b013b62776f92472e7f9df8608

GitOrigin-RevId: 0d481c684eb2ad4d0e54c307407b4d3635f18277