Unify project reloading
across different actions.

This is likely to be a temporary state until we come with a better
structure  that works for project loading, sync, building dependencies,
building runtime artifacts etc. For now, the goal is to make make all
cases use the same primitives so that the structure can be re-arranged
as it works better.

Bug: 417935665
Test: n/a
Change-Id: I9fbe0cc9d0ee69c2deb88b800f479e4832b9f6c3

GitOrigin-RevId: bf612e8f850e026feb97a4b04c0a33928c1bce62