Update project snapshot incrementally
This is to apply project structure changes at the end of sync operation
automatically and only if there are changes.

Unfortunately it means relying on global variables like the
`SnapshotHolder` even more. We might be able to pass the state
explicitly later but it is difficult with IntelliJ itself being built
around the service locator pattern.

Bug: 417935665
Test: n/a
Change-Id: Ia838b77e0be714206b12a4881991482513bd6201

GitOrigin-RevId: d985261b5154da7b79badb95160843ac0c73cbda